chore(examples): updated the example of with-react-jss to utlize the App router
