+++ examples/with-react-jss/app/layout.tsx
@@ -1,0 +1,15 @@
+import StyledJsxRegistry from "../lib/registry";
+
+export default function RootLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return (
+    <html lang="en">
+      <body>
+        <StyledJsxRegistry>{children}</StyledJsxRegistry>
+      </body>
+    </html>
+  );
+}
