+++ examples/with-react-jss/package.json
@@ -7,13 +7,13 @@
   },
   "dependencies": {
     "next": "latest",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "react-jss": "^10.9.2"
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
+    "react-jss": "^10.10.0"
   },
   "devDependencies": {
-    "@types/node": "18.7.15",
-    "@types/react": "18.2.8",
-    "typescript": "4.8.2"
+    "@types/node": "^22.10.2",
+    "@types/react": "^19.0.2",
+    "typescript": "^5.7.2"
   }
 }
